Toolbar renaming from north-header and move styles to component
