+++ admin/client/src/components/breadcrumb/README.md
@@ -1,4 +1,4 @@
-#NorthHeaderBreadcrumbs
+#ToolbarBreadcrumbs
 
 The breadcrumbs for the current section of the CMS.
