--- admin/client/src/components/north-header/styles.scss
@@ -1,32 +1,0 @@
-.north-header {
-    border-bottom: 1px solid $border-color-dark;
-    background-color: $background-north;
-    width: 100%;
-    height: $toolbar-total-height;
-    z-index: 60;
-}
-
-// Typically to hold breadcrumbs and back button
-.north-header__navigation {
-    float: left;
-    padding: 0 $spacer-y;
-    max-width: 70%;
-    max-height: $toolbar-height;
-}
-
-.toolbar__navigation__back-button {
-  float: left;
-  margin: $spacer-y/2 $spacer-x $spacer-y/2 0;
-}
-
-.north-header__heading {
-    font-size: $font-size-lg;
-    font-weight: normal;
-    line-height: 20px;
-    margin: $spacer-y 0;
-}
-
-// To hold things like search icon and view toggle
-.north-header__actions {
-
-}
