+++ admin/client/src/components/toolbar/README.md
@@ -1,4 +1,4 @@
-#NorthHeader
+#Toolbar
 
 The main header for sections in the CMS.
 Styles are contained within layout.scss
