+++ admin/client/src/components/toolbar/styles.scss
@@ -1,0 +1,46 @@
+.toolbar,
+.toolbar--north,
+.toolbar--content,
+.toolbar--south {
+  @include make-container();
+  height: $toolbar-total-height;
+  background-color: $body-bg;
+
+  .btn-toolbar {
+    // TODO Remove '.cms .btn-toolbar' override
+    margin-top: 0;
+    margin-bottom: 0;
+  }
+}
+
+// Northern bar containing breadcrum, tabs etc.
+.toolbar--north {
+  display: block;
+  background-image: none;
+  background-color: $background-north;
+  border-bottom: 1px solid $border-color;
+  z-index: 60;
+}
+
+// Secondary content actions eg. Add page button
+.toolbar--content {
+  padding: $spacer-y*.625 $spacer-x*.75;
+  margin-bottom: $spacer-y/2;
+}
+
+// Primary content actions
+.toolbar--south {
+  padding-top: $spacer-y*.625;
+  padding-bottom: $spacer-y*.625;
+  position: absolute;
+  bottom: 0;
+  border-top: 1px solid $border-color;
+}
+
+.toolbar__back-button {
+  float: left;
+  margin: $spacer-y*.625 $spacer-x/2 $spacer-y*.625 #{-$spacer-x/4};
+  padding-left: $spacer-x/4;
+  padding-right: $spacer-x/4;
+  border: 0;
+}
