+++ admin/client/src/components/toolbar/tests/toolbar-test.js
@@ -2,6 +2,6 @@
 
 jest.dontMock('../index');
 
-describe('NorthHeaderComponent', () => {
+describe('ToolbarComponent', () => {
 
 });
