+++ admin/client/src/styles/_layout.scss
@@ -53,45 +53,3 @@
     }
   }
 }
-
-
-// TOOLBARS
-.toolbar {
-	width: 100%;        						// Todo: replace with mixin @include make-container();
-	height: $toolbar-total-height;
-	background-color: $body-bg;
-
- 	.btn-toolbar {
-		// TODO Remove '.cms .btn-toolbar' override
-		margin-top: 0;
-		margin-bottom: 0;
-	}
-}
-
-// Northern bar containing breadcrum, tabs etc.
-.toolbar--north {
-	@extend .toolbar;
-	padding-left: $spacer-x;				// Todo: replace with mixin @include make-container();
-	padding-right: $spacer-x;				// Todo: replace with mixin @include make-container();
-	display: block;
-	background-image: none;
-	background-color: $background-north;
-	border-bottom: 1px solid $border-color;
-	z-index: 60;
-}
-
-// Secondary content actions eg. Add page button
-.toolbar--content {
-	@extend .toolbar;
-	padding: $spacer-y*.625 $spacer-x*.75;
-	margin-bottom: $spacer-y/2;
-}
-
-// Primary content actions
-.toolbar--south {
-	@extend .toolbar;
-	padding: $spacer-y*.625 $spacer-x;
-	position: absolute;
-	bottom: 0;
-	border-top: 1px solid $border-color;
-}
