+++ admin/client/src/styles/bundle.scss
@@ -34,7 +34,7 @@
 @import "../components/hidden-field/styles";
 @import "../components/label/styles";
 @import "../components/preview/styles";
-@import "../components/north-header/styles";
+@import "../components/toolbar/styles";
 @import "../components/form/styles";
 
 // Layout and sections
